Removing cursed line for the third time...
This line was added as tests in fab4bb7b2264ff14d6011adbadd1ee484a712146 and 1c7f428a842364246c6b6c6e1f45781acb1d85ca. But removed in 3978ebd5006f6ddb4b3325c29f4129f065623591. We dont understand how it came back. Is in master and tag 0.8. But is not in the build release of 0.8 we did...